requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2
passlib[bcrypt]
python-jose[cryptography]
python-multipart